fix: "Режиссёры" внесены корректировки по 1 ревью.

+++ src/main/java/ru/yandex/practicum/filmorate/model/FilmDirector.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class FilmDirector {
     private Long filmId;
-    private Integer directorId;
+    private Long directorId;
 }
